test: move format_docstring tests to new file and update to use pytest
